Refactor `extract_new_rows()` into `utils`
